+++ src/prefect/indicators/run.py
@@ -22,3 +22,4 @@ from .usage.u10 import u10
 from .usage.u11 import u11
 from .usage.u12 import u12
 from .usage.u13 import u13
+from .usage.u14 import u14