+++ src/prefect/indicators/usage/u5.py
@@ -35,7 +35,6 @@ SELECT
 FROM
     Session
     INNER JOIN statique ON point_de_charge_id = pdc_id
-    LEFT JOIN City ON City.code = code_insee_commune
     $join_extras
 WHERE
     $timespan