+++ src/prefect/tests/infrastructure/test_i4.py
@@ -18,10 +18,10 @@ from ..parameters import (
     PARAMETERS_CHUNK,
 )
 
-# expected result for level [city, epci, dpt, reg]
-N_LEVEL = [8, 490, 849, 2436]
+# expected result for level [city, epci, dpt, reg, operationalunit]
+N_LEVEL = [8, 490, 849, 2436, 664]
 N_DPTS = 109
-N_NAT_REG_DPT_EPCI_CITY = 36465
+N_NAT_REG_DPT_EPCI_CITY_OU = 36984
 
 TIMESPAN = IndicatorTimeSpan(start=datetime.now(), period=IndicatorPeriod.DAY)
 PARAMETERS_FLOW = [prm + (lvl,) for prm, lvl in zip(PARAM_FLOW, N_LEVEL, strict=True)]
@@ -73,13 +73,14 @@ def test_flow_i4_national(db_connection):
 
 def test_flow_i4(db_connection):
     """Test the `i4` flow."""
-    expected = N_NAT_REG_DPT_EPCI_CITY
+    expected = N_NAT_REG_DPT_EPCI_CITY_OU
     all_levels = [
         Level.NATIONAL,
         Level.REGION,
         Level.DEPARTMENT,
         Level.CITY,
         Level.EPCI,
+        Level.OPERATIONALUNIT,
     ]
     indicators = i4.i4(
         Environment.TEST,